+++ src/components/BackgroundLayer.js
@@ -342,12 +342,14 @@ export class BackgroundLayer extends React.Component {
 
   render() {
     return (
-      <div className="rct9k-background-layer">
-        {this.renderHighlightedWeekends()}
-        {this.renderCustomComponents(this.props.highlightedIntervals)}
-        {this.renderNowMarker()}
-        {this.renderCustomComponents(this.props.markers)}
-        {this.renderVerticalGrid()}
+      <div className="rct9k-background-layer-wrapper">
+        <div className="rct9k-background-layer">
+          {this.renderHighlightedWeekends()}
+          {this.renderCustomComponents(this.props.highlightedIntervals)}
+          {this.renderNowMarker()}
+          {this.renderCustomComponents(this.props.markers)}
+          {this.renderVerticalGrid()}
+        </div>
       </div>
     );
   }